+++ zombiehero/src/main/kotlin/click/recraft/zombiehero/player/WalkSpeed.kt
@@ -1,5 +1,5 @@
 package click.recraft.zombiehero.player
 
 interface WalkSpeed {
-    val walkSpeed: Float
+    val walkSpeed: Int
 }
